requirements.txt
python-dotenv
pymongo
certifi
selenium